requirements.txt
langchain
python-dotenv
llama-index>=0.9.31 
pinecone-client>=3.0.0
llama-index-llms-gemini
google-generativeai
llama-index-vector-stores-pinecone
llama-index-embeddings-gemini
torch 
transformers 
python-pptx 
Pillow
# -e.